Fixed some bugs
In multiple files `HashMap<String, Integer>` was mistakenly used instead of `HashMap<String, Double>`

+++ Ilya-Dudnikov/term_2/block_3/task_2/src/main/java/model/WeatherData/WeatherData.java
@@ -1,12 +1,10 @@
 package model.WeatherData;
 
-import java.util.Objects;
-
 public record WeatherData(
-		int airTemperature,
-		int cloudCover,
-		int humidity,
-		int precipitation,
-		int windDirection,
-		int windSpeed) {
+		Double airTemperature,
+		Double cloudCover,
+		Double humidity,
+		Double precipitation,
+		Double windDirection,
+		Double windSpeed) {
 }
