+++ Ilya-Dudnikov/term_2/block_3/task_2/src/main/java/model/WeatherModel/OpenWeatherMapModel.java
@@ -6,7 +6,7 @@ import java.util.Map;
 
 public class OpenWeatherMapModel extends WeatherModel {
 	@Override
-	public void updateState(Map<String, Integer> newStats) {
+	public void updateState(Map<String, Double> newStats) {
 		data = new WeatherData(
 				newStats.get("airTemperature"),
 				newStats.get("cloudCover"),
