+++ Ilya-Dudnikov/term_2/block_3/task_2/src/main/java/model/WeatherModel/StormglassModel.java
@@ -6,7 +6,7 @@ import java.util.Map;
 
 public class StormglassModel extends WeatherModel {
 	@Override
-	public void updateState(Map<String, Integer> newStats) {
+	public void updateState(Map<String, Double> newStats) {
 		data = new WeatherData(
 				newStats.get("airTemperature"),
 				newStats.get("cloudCover"),
